+++ src/engine/Screen.ts
@@ -439,8 +439,17 @@ export class Screen {
   /**
    * Requests to go fullscreen using the browser fullscreen api, requires user interaction to be successful.
    * For example, wire this to a user click handler.
+   *
+   * Optionally specify a target element id to go fullscreen, by default the game canvas is used
+   * @param elementId
    */
-  public goFullScreen(): Promise<void> {
+  public goFullScreen(elementId?: string): Promise<void> {
+    if (elementId) {
+      const maybeElement = document.getElementById(elementId);
+      if (maybeElement) {
+        return maybeElement.requestFullscreen();
+      }
+    }
     return this._canvas.requestFullscreen();
   }